refactor: Change money and rent to int, add game config options
This commit refactors the game to use integers for money and rent values instead of decimals.
This change was made to simplify calculations and avoid potential rounding errors.

Additionally, this commit adds several new configuration options to the GameConfig model,
allowing for greater customization of game rules.

+++ src/Hubs/GameHubs.cs
@@ -168,12 +168,12 @@ public class GameHubs : Hub<IResponse>
     }
 
     // Trade stuff     
-    public async Task InitiateTrade(Guid gameId, Guid recipientId, List<Guid> propertyOffer, List<Guid> propertyCounterOffer, decimal moneyFromInitiator, decimal moneyFromRecipient)
+    public async Task InitiateTrade(Guid gameId, Guid recipientId, List<Guid> propertyOffer, List<Guid> propertyCounterOffer, int moneyFromInitiator, int moneyFromRecipient)
     {
         var initiatorId = GetPlayerId();
         await _gameService.InitiateTrade(gameId, initiatorId, recipientId, propertyOffer, propertyCounterOffer, moneyFromInitiator, moneyFromRecipient);
     }
-    public async Task NegotiateTrade(Guid gameId,Guid tradeId, List<Guid> propertyOffer, List<Guid> propertyCounterOffer, decimal moneyFromInitiator, decimal moneyFromRecipient)
+    public async Task NegotiateTrade(Guid gameId,Guid tradeId, List<Guid> propertyOffer, List<Guid> propertyCounterOffer, int moneyFromInitiator, int moneyFromRecipient)
     {
         var negotiatorId = GetPlayerId();
         await _gameService.NegotiateTrade(gameId, negotiatorId, tradeId, propertyOffer, propertyCounterOffer, moneyFromInitiator, moneyFromRecipient);
